fix deserialize method: - remove cartid in the method - calculate the total by item - assign updated time by the server side datetime.now() - add test to current unittest
tests items related function(calling the serialize and deserialize functions) using thunderclients:
- addItem()
- deleteItem()
- updateItem()

+++ service/routes.py
@@ -63,9 +63,10 @@ def create_items(old_cart_id):
     # Create an item from the json data
     item = Item()
     item.deserialize(request.get_json())
-
+    item.shopcart_id = old_cart_id
     # Append the item to the shopcart
     shopcart.items.append(item)
+    shopcart.total_price = shopcart.get_total_price()
     shopcart.update()
 
     # Prepare a message to return
@@ -91,12 +92,26 @@ def delete_items(shopcart_id, item_id):
         raise TypeError("item_id should be int")
     check_content_type("application/json")
 
+    shopcart = Shopcart.find(shopcart_id)
+    if not shopcart:
+        abort(
+            status.HTTP_404_NOT_FOUND,
+            f"Shopcart with id '{shopcart_id}' could not be found.",
+        )
+
     # See if the item exists and delete it if it does
     item = Item.find(item_id)
+    if not item:
+        abort(status.HTTP_404_NOT_FOUND, f"Item with id '{item_id}' was not found.")
+
     if item:
+        shopcart.items.remove(item)
         item.delete()
 
-    app.logger.info("Item with ID [%s] deleted.", item.id)
+    shopcart.total_price = shopcart.get_total_price()
+    shopcart.update()
+
+    app.logger.info("Item with ID [%s] deleted.", item_id)
 
     return make_response("", status.HTTP_204_NO_CONTENT)
 
@@ -161,7 +176,8 @@ def update_item(cart_id, item_id):
 
     item.id = item_id
     item.update()
-
+    cart.total_price = cart.get_total_price()
+    cart.update()
     app.logger.info("Item with ID [%s] updated.", item.id)
     return jsonify(item.serialize()), status.HTTP_200_OK
 
@@ -204,11 +220,13 @@ def create_shopcarts():
     Creates a Shopcart
     This endpoint will create an Shopcart based the data in the body that is posted
     """
+    print("create shopcart")
     app.logger.info("Request to create an Shopcart")
     check_content_type("application/json")
     # Create the shopcart
     shopcart = Shopcart()
     shopcart.deserialize(request.get_json())
+
     shopcart.create()
     # Create a message to return
     message = shopcart.serialize()
@@ -349,13 +367,9 @@ def list_shopcarts():
     min_price = request.args.get("minprice")
 
     if max_price:
-        results = [
-            cart for cart in results if cart["total_price"] <= float(max_price)
-        ]
+        results = [cart for cart in results if cart["total_price"] <= float(max_price)]
     if min_price:
-        results = [
-            cart for cart in results if cart["total_price"] >= float(min_price)
-        ]
+        results = [cart for cart in results if cart["total_price"] >= float(min_price)]
     if item_id:
         temp = []
         for cart in results:
